requirements.txt
Django
Pillow
MySQL-python
django-crispy-forms==1.4.0
python-dateutil
